+++ Web/Scotty/Trans.hs
@@ -39,6 +39,7 @@ module Web.Scotty.Trans
     , getResponseHeaders, getResponseStatus, getResponseContent
       -- ** Exceptions
     , raise, raiseStatus, throw, rescue, next, finish, defaultHandler, liftAndCatchIO
+    , liftIO, catch
     , StatusError(..)
       -- * Parsing Parameters
     , Param, Parsable(..), readEither
@@ -66,7 +67,7 @@ import Web.Scotty.Route
 import Web.Scotty.Internal.Types (ActionT(..), ScottyT(..), defaultScottyState, Application, RoutePattern, Options(..), defaultOptions, RouteOptions(..), defaultRouteOptions, ErrorHandler, Kilobytes, File, addMiddleware, setHandler, updateMaxRequestBodySize, routes, middlewares, ScottyException(..), ScottyState, defaultScottyState, StatusError(..), Content(..))
 import Web.Scotty.Util (socketDescription)
 import Web.Scotty.Body (newBodyInfo)
-import UnliftIO.Exception (Handler(..), catches)
+import UnliftIO.Exception (Handler(..), catch, catches)
 
 -- | Run a scotty application using the warp server.
 -- NB: scotty p === scottyT p id